NP-9089 Extract common fields as separate components
